Implement the compilation feature to native machine code

See merge request sherifnasser/AinLanguage!2

+++ src/expression/value/DoubleValue.cpp
@@ -1,7 +1,7 @@
 #include "DoubleValue.hpp"
 #include "Type.hpp"
 
-DoubleValue::DoubleValue(long double value)
+DoubleValue::DoubleValue(double value)
 :IValue(Type::DOUBLE),PrimitiveValue(Type::DOUBLE,value){}
 
 std::wstring DoubleValue::toString(){
